fix bug --do separation module ( no helper )
