+++ rs-son/src/components/CoachPage/CoachDetail.vue
@@ -1,8 +1,8 @@
-// FIXED 
+// FIXED
 <template>
   <li>
     <h3>{{ itemList[1].firstName }}</h3>
-    <h4>${{ itemList[1].hourlyRate }}/{{$t('hour')}}</h4>
+    <h4>${{ itemList[1].hourlyRate }}/{{ $t("hour") }}</h4>
     <div>
       <item-badge
         v-for="(itemArea, index) in itemList[1].areas"
@@ -20,7 +20,7 @@
         }"
         class="outline"
       >
-        {{$t('contact')}}
+        {{ $t("contact") }}
       </item-link>
       <item-link
         :linkTo="{
@@ -28,18 +28,14 @@
           params: { id: itemList[0] },
         }"
       >
-        {{$t('viewDetail')}}
+        {{ $t("viewDetail") }}
       </item-link>
     </div>
   </li>
 </template>
 
 <script>
-import ItemBadge from "../Common/ItemBadge";
-import ItemLink from "../Common/ItemLink";
-
 export default {
-  components: { ItemLink, ItemBadge },
   props: ["itemList"],
   data() {
     return {
