+++ rs-son/src/store/index.js
@@ -2,184 +2,14 @@
 import {
   createStore
 } from 'vuex'
-import axios from "axios";
+import auth from './modules/auth'
+import coach from './modules/coach'
 
-import * as firebase from '../env'
-
-export default createStore({
-  state: {
-    areas: ["frontend", "backend", "career"],
-    coaches: [],
-    coachesTemp: [],
-    dataDetail: [],
-    tokenId: JSON.parse(localStorage.getItem('checkLogin')),
-    linkTo: '/coaches',
-    status: '',
-    request: [],
-    checkLogin: true,
-    loading: false,
-    locale: localStorage.getItem("lang") ? localStorage.getItem("lang") : "gb",
-  },
-  getters: {
-    getLinkTo: (state) => {
-      if (state.tokenId != null && state.tokenId != '') {
-        state.linkTo = '/register'
-      } else {
-        state.linkTo = {
-          path: '/auth',
-          query: {
-            redirect: 'register'
-          },
-          name: 'Auth'
-        }
-      }
-      return state.linkTo;
-    },
-  },
-  mutations: {
-    SET_LOCALE(state, locale) {
-      return state.locale = locale
-    },
-    SET_DEFAULT_DATA(state, coaches) {
-      return state.coaches = coaches;
-    },
-    SET_TEMP_DATA(state, coachesTemp) {
-      return state.coachesTemp = coachesTemp;
-    },
-    SET_DATA_FILTER(state, areas) {
-      return state.areas = areas;
-    },
-    SET_DATA_REQUEST(state, request) {
-      return state.request = request;
-    },
-    SET_DATA_DETAIL(state, dataDetail) {
-      return state.dataDetail = dataDetail;
-    },
-    SET_RESULT_POST(state, status) {
-      return state.status = status;
-    },
-    SET_CHECK_LOGIN(state, checkLogin) {
-      return state.checkLogin = checkLogin;
-    },
-    SET_TOKEN_ID(state, tokenId) {
-      return state.tokenId = tokenId;
-    },
-    SET_LOADING(state, loading) {
-      return state.loading = loading;
-    },
-  },
-  actions: {
-    // GET DATA COACHES DEFAUT FROM API
-    getDefaultData({
-      commit
-    }) {
-      axios
-        .get("https://coaches-project-8d77f-default-rtdb.firebaseio.com/coaches.json")
-        .then((res) => {
-          commit('SET_DEFAULT_DATA', res.data);
-          commit('SET_TEMP_DATA', res.data)
-          commit('SET_LOADING', false);
-        }).catch(err => console.log(err));
-    },
-    getDataDetail({
-      commit
-    }, payLoad) {
-      axios
-        .get(payLoad.url)
-        .then((res) => {
-          commit('SET_DATA_DETAIL', res.data);
-        }).catch(err => console.log(err));
-    },
-    // GET DATA REQUEST FOR MEM FROM API
-    getDataRequest(store) {
-      let userId = store.state.tokenId.localId;
-      axios
-        .get(`${firebase.API_DATA_JSON}/request/${userId}.json`)
-        .then((res) => {
-          store.commit('SET_DATA_REQUEST', res.data);
-          store.commit('SET_LOADING', false);
-          console.log(res.data);
-        }).catch(err => console.log(err));
-    },
-    // GET DATA TEMP TO SS WITH FILTER
-    getDatafilter(store, payLoad) {
-      store.commit('SET_DATA_FILTER', payLoad.listFilter)
-      let temp = Object.values(store.state.coachesTemp).filter((item) => {
-        let check;
-        payLoad.listFilter.forEach(elm => {
-          if (item.areas.indexOf(elm) != -1) {
-            check = true
-          }
-        })
-        return check;
-      })
-      store.state.coaches = temp;
-    },
-    // HANDLE REGISTER COACH
-    handlePostDataCoach({
-      commit
-    }, payLoad) {
-      axios.put(`${firebase.API_DATA_JSON}/coaches/${payLoad.user.localId}.json?auth=${payLoad.user.idToken}`, payLoad.data).then((res) => {
-        console.log('POST CORRECTED');
-        commit('SET_RESULT_POST', res.data);
-      }).catch(err => {
-        console.log('POST FAILSE');
-        commit('SET_RESULT_POST', err.message)
-      })
-    },
-    // HANDLE POST REQUEST FROM USER
-    handlePostDataRequest(store, payLoad) {
-      axios.post(`${firebase.API_DATA_JSON}/request/${payLoad.requestId}.json`, payLoad.data).then((res) => {
-        console.log('POST REQUESTS CORRECTED');
-        console.log(res);
-      }).catch(err => {
-        console.log('POST REQUESTS FAILSE');
-        console.log(err);
-      })
-    },
-    // HANDLE SIGN FOR USER
-    async handleSignUp({
-      commit
-    }, payLoad) {
-      commit('SET_LOADING', true);
-      setTimeout(() => {
-        commit('SET_LOADING', false);
-      }, 500)
-      try {
-        let res = await axios.post(`${firebase.API_AUTH}:signUp?key=${firebase.API_KEY}`, payLoad.data);
-        commit('SET_LOADING', false);
-        commit('SET_TOKEN_ID', res.data);
-        localStorage.setItem("checkLogin", JSON.stringify(res.data));
-        localStorage.setItem("userId", JSON.stringify(res.data));
-        return true
-      } catch (err) {
-        console.log(err);
-        return false;
-      }
-    },
-    // HANDLE LOGIN FOR MEM
-    async handleLogin({
-      commit,
-      // state
-    }, payLoad) {
-      commit('SET_LOADING', true);
-      setTimeout(() => {
-        commit('SET_LOADING', false);
-      }, 500)
-      try {
-        let res = await axios.post(`${firebase.API_AUTH}:signInWithPassword?key=${firebase.API_KEY}`, payLoad.data)
-        let checkLogin = {
-          idToken: res.data.idToken,
-          localId: res.data.localId,
-          email: res.data.email,
-        };
-        commit('SET_TOKEN_ID', checkLogin);
-        localStorage.setItem("checkLogin", JSON.stringify(checkLogin));
-        return true;
-      } catch (err) {
-        console.log(err);
-        return false;
-      }
-    },
+const store = createStore({
+  modules: {
+    auth: auth,
+    coach: coach
   },
 })
+
+export default store
