+++ rs-son/src/views/Auth.vue
@@ -7,10 +7,9 @@
 </template>
 
 <script>
-import ItemCard from "../components/Common/ItemCard";
 import ItemFormAuth from "../components/AuthPage/ItemFormAuth.vue";
 export default {
-  components: { ItemCard, ItemFormAuth },
+  components: { ItemFormAuth },
   beforeRouteLeave() {
     console.log("GOODBYE");
   },
